+++ src/proxy/src/logic/group_logic.rs
@@ -17,7 +17,7 @@ use crate::{
         RewardBufferStore, StorageInsertable, StorageInsertableByKey, StorageQueryable,
         StorageUpdateable,
     },
-    MAX_GROUPS_PER_USER,
+    USER_GROUP_CREATION_LIMIT,
 };
 use candid::Principal;
 use canister_types::{
@@ -77,9 +77,10 @@ impl GroupCalls {
         // Get the member and add the group to the member
         let (_, mut member) = MemberStore::get(caller())?;
 
-        if member.get_owned().len() >= MAX_GROUPS_PER_USER {
-            return Err(ApiError::bad_request()
-                .add_message(format!("You can only own {} groups", MAX_GROUPS_PER_USER).as_str()));
+        if member.get_owned().len() >= USER_GROUP_CREATION_LIMIT {
+            return Err(ApiError::bad_request().add_message(
+                format!("You can only own {} groups", USER_GROUP_CREATION_LIMIT).as_str(),
+            ));
         }
 
         // Create and store the group