+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#[Unreleased]
+
+### Changed
+
+- MAX_GROUPS_PER_USER to USER_GROUP_CREATION_LIMIT
+
 ## [0.2.2]
 
 ### Added
 
-- limit of 10 groups per user
+- limit creating groups to 10 groups per user
 
 ### Changed